Revert "feat: add basic frontend bundler setup"
