--- assets/shared.98e1cdfe.js
@@ -1,5 +1,0 @@
-import { updateCartCount } from './cart.f78a7330.js';
-
-document.addEventListener('DOMContentLoaded', () => {
-  updateCartCount();
-});
