--- frontend/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "ultrablack-frontend",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "dependencies": {
-    "@fontsource/inter": "^5.0.8",
-    "gsap": "^3.12.2",
-    "swiper": "^11.0.5"
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.16",
-    "postcss": "^8.4.32",
-    "tailwindcss": "^3.4.1",
-    "vite": "^5.2.0"
-  }
-}
