--- frontend/src/main.css
@@ -1,4 +1,0 @@
-/* Placeholder for Tailwind build */
-body {
-  font-family: sans-serif;
-}
