--- frontend/src/shared.js
@@ -1,5 +1,0 @@
-import { updateCartCount } from './cart.js';
-
-document.addEventListener('DOMContentLoaded', () => {
-  updateCartCount();
-});
