--- frontend/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ['../*.html', './src/**/*.{js,ts}'],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
