+++ index.html
@@ -58,7 +58,18 @@
     }
     </script>
     
-    <link rel="stylesheet" href="assets/main.86ff73c8.css">
+    <!-- Tailwind CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    
+    <!-- Google Fonts: Inter -->
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;900&display=swap" rel="stylesheet">
+    
+    <!-- GSAP (GreenSock Animation Platform) -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/TextPlugin.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js"></script>
 
     <style>
         /* Custom Styles */
@@ -270,7 +281,51 @@
        </div>
    </footer>
 
-        <script type="module" src="assets/homepage.c706e55e.js"></script>
+        <script src="js/cart.js" defer></script>
+        <script src="js/homepage.js" defer></script>
+    <script>
+        document.addEventListener('DOMContentLoaded', function() {
+            gsap.registerPlugin(ScrollTrigger);
+
+            // Mobile Menu Toggle
+            const mobileMenuButton = document.getElementById('mobile-menu-button');
+            const mobileMenu = document.getElementById('mobile-menu');
+            mobileMenuButton.addEventListener('click', () => {
+                mobileMenu.classList.toggle('hidden');
+            });
+            
+            // Initialize cart count
+           updateCartCount();
+
+            // GSAP Animation for the core statement
+            const animatedText = document.getElementById('animated-text');
+            const textContent = "We run together. We move the culture.";
+            
+            const words = textContent.split(' ');
+            animatedText.innerHTML = words.map(word => `<span class="inline-block opacity-0">${word}</span>`).join(' ');
+            
+            gsap.to("#animated-text span", {
+                duration: 0.8,
+                opacity: 1,
+                y: 0,
+                ease: "power3.out",
+                stagger: 0.15,
+                scrollTrigger: {
+                    trigger: "#animated-text",
+                    start: "top 80%",
+                }
+            });
+
+            // Simple parallax for hero background on scroll
+            window.addEventListener('scroll', () => {
+                const heroBg = document.querySelector('.hero-bg');
+                const scrollY = window.scrollY;
+                if (heroBg) {
+                    heroBg.style.transform = `translateY(${scrollY * 0.3}px)`;
+                }
+            });
+        });
+    </script>
 
 </body>
 </html>
